pyproject.toml
[tool.poetry]
name = "cratd2c-stakingmanager-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"
marshmallow = "^3.21.3"
marshmallow-dataclass = "^8.7.0"
marshmallow-enum = "^1.5.1"
pyyaml = "^6.0.1"
django-extensions = "^3.2.3"
ptpython = "^3.0.27"
pytest-django = "^4.8.0"
mixer = "^7.2.2"
web3 = "^6.20.0"
redis = "^5.0.7"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
gunicorn = "^22.0.0"
pillow = "^10.4.0"
django-filter = "^24.2"
drf-extra-fields = "^3.7.0"
shyaml = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
